Fix doc tests
